move long-running forensics scans to separate service, throttle backend calls
