+++ backend/mempool-config.sample.json
@@ -82,7 +82,8 @@
     "BACKEND": "lnd",
     "STATS_REFRESH_INTERVAL": 600,
     "GRAPH_REFRESH_INTERVAL": 600,
-    "LOGGER_UPDATE_INTERVAL": 30
+    "LOGGER_UPDATE_INTERVAL": 30,
+    "FORENSICS_INTERVAL": 43200
   },
   "LND": {
     "TLS_CERT_PATH": "tls.cert",
