+++ backend/src/__fixtures__/mempool-config.template.json
@@ -98,7 +98,8 @@
     "TOPOLOGY_FOLDER": "__LIGHTNING_TOPOLOGY_FOLDER__",
     "STATS_REFRESH_INTERVAL": 600,
     "GRAPH_REFRESH_INTERVAL": 600,
-    "LOGGER_UPDATE_INTERVAL": 30
+    "LOGGER_UPDATE_INTERVAL": 30,
+    "FORENSICS_INTERVAL": 43200
   },
   "LND": {
     "TLS_CERT_PATH": "",
